Refactor Home into progressively streamed sections
Split the Home page so the header renders immediately while the heavy
Boards and Notes sections stream in independently via Suspense.

- page.tsx now only awaits cheap header data (boards count + latest image)
  and renders BoardsSection/NotesSection inside their own Suspense boundaries
- BoardsSection/NotesSection are async server components that fetch and
  filter their own data
- HomeTabs split into client presentational BoardsGrid and NotesGrid
- Add BoardsSkeleton/NotesSkeleton fallbacks matching the grid layouts
